aws_de/tests/requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
boto3==1.28.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
confluent-kafka==2.3.0
pyspark==3.4.1
python-dateutil==2.8.2

